refactor : User 도메인 코드 정리, Json 요청 응답 형식 변경
