+++ src/main/java/com/dss/realworld/user/api/UserController.java
@@ -1,8 +1,9 @@
 package com.dss.realworld.user.api;
 
 import com.dss.realworld.user.api.dto.AddUserRequestDto;
-import com.dss.realworld.user.api.dto.UserResponseDto;
+import com.dss.realworld.user.api.dto.LoginUserRequestDto;
 import com.dss.realworld.user.api.dto.UpdateUserRequestDto;
+import com.dss.realworld.user.api.dto.UserResponseDto;
 import com.dss.realworld.user.app.UserService;
 import com.dss.realworld.user.domain.User;
 import lombok.RequiredArgsConstructor;
@@ -29,14 +30,16 @@ public class UserController {
         return new UserResponseDto(user);
     }
 
-    @PostMapping(value = "users/login")
-    public AddUserResponseDto login(@RequestBody LoginUserRequestDto loginUserRequestDto) {
+    @PostMapping(value = "/users/login")
+    public UserResponseDto login(@RequestBody LoginUserRequestDto loginUserRequestDto) {
         User user = userService.login(loginUserRequestDto);
 
-        return new AddUserResponseDto(user);
+        return new UserResponseDto(user);
     }
 
     private Long getLoginUserId() {
         return 1L;
     }
 }
+
+//todo ResponseEntity를 반환하도록 수정
