+++ src/main/java/com/dss/realworld/user/api/dto/AddUserRequestDto.java
@@ -1,11 +1,15 @@
 package com.dss.realworld.user.api.dto;
 
-import com.fasterxml.jackson.annotation.JsonRootName;
+import com.fasterxml.jackson.annotation.JsonTypeInfo;
+import com.fasterxml.jackson.annotation.JsonTypeInfo.As;
+import com.fasterxml.jackson.annotation.JsonTypeInfo.Id;
+import com.fasterxml.jackson.annotation.JsonTypeName;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@JsonRootName(value = "user")
+@JsonTypeName(value = "user")
+@JsonTypeInfo(include = As.WRAPPER_OBJECT, use = Id.NAME)
 @Getter
 @NoArgsConstructor
 public class AddUserRequestDto {
